requirements.txt
asgiref==3.7.2
crispy-tailwind==0.5.0
Django==4.2.4
django-crispy-forms==2.0
django-environ==0.10.0
export==0.2.1
psycopg2==2.9.7
psycopg2-binary==2.9.7
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
whitenoise==6.5.0
